Pull all functionality from NOD controllers into a concern (so v2 can be exactly the same as v1 with the exception of the api_version)
