+++ modules/appeals_api/app/controllers/appeals_api/v1/decision_reviews/notice_of_disagreements_controller.rb
@@ -6,121 +6,11 @@ require 'appeals_api/form_schemas'
 class AppealsApi::V1::DecisionReviews::NoticeOfDisagreementsController < AppealsApi::ApplicationController
   include AppealsApi::JsonFormatValidation
   include AppealsApi::StatusSimulation
-
-  skip_before_action(:authenticate)
-  before_action :validate_json_format, if: -> { request.post? }
-  before_action :validate_json_schema, only: %i[create validate]
-  before_action :new_notice_of_disagreement, only: %i[create validate]
-  before_action :find_notice_of_disagreement, only: %i[show]
-
-  FORM_NUMBER = '10182'
-  MODEL_ERROR_STATUS = 422
-  HEADERS = JSON.parse(
-    File.read(
-      AppealsApi::Engine.root.join('config/schemas/v1/10182_headers.json')
-    )
-  )['definitions']['nodCreateHeadersRoot']['properties'].keys
-  SCHEMA_ERROR_TYPE = Common::Exceptions::DetailedSchemaErrors
-
-  def create
-    @notice_of_disagreement.save
-    AppealsApi::NoticeOfDisagreementPdfSubmitJob.perform_async(@notice_of_disagreement.id)
-    render_notice_of_disagreement
-  end
-
-  def show
-    @notice_of_disagreement = with_status_simulation(@notice_of_disagreement) if status_requested_and_allowed?
-    render_notice_of_disagreement
-  end
-
-  def validate
-    render json: validation_success
-  end
-
-  def schema
-    render json: AppealsApi::JsonSchemaToSwaggerConverter.remove_comments(
-      AppealsApi::FormSchemas.new.schema(self.class::FORM_NUMBER)
-    )
-  end
+  include AppealsApi::NodV1
 
   private
 
-  def validate_json_schema
-    validate_json_schema_for_headers
-    validate_json_schema_for_body
-  rescue SCHEMA_ERROR_TYPE => e
-    render json: { errors: e.errors }, status: 422
-  end
-
-  def validate_json_schema_for_headers
-    AppealsApi::FormSchemas.new(SCHEMA_ERROR_TYPE).validate!("#{FORM_NUMBER}_HEADERS", request_headers)
-  end
-
-  def validate_json_schema_for_body
-    AppealsApi::FormSchemas.new(SCHEMA_ERROR_TYPE).validate!(FORM_NUMBER, @json_body)
-  end
-
-  def validation_success
-    {
-      data: {
-        type: 'noticeOfDisagreementValidation',
-        attributes: {
-          status: 'valid'
-        }
-      }
-    }
-  end
-
-  def request_headers
-    HEADERS.reduce({}) do |hash, key|
-      hash.merge(key => request.headers[key])
-    end.compact
-  end
-
-  def new_notice_of_disagreement
-    @notice_of_disagreement = AppealsApi::NoticeOfDisagreement.new(
-      auth_headers: request_headers,
-      form_data: @json_body,
-      source: request_headers['X-Consumer-Username'],
-      board_review_option: @json_body['data']['attributes']['boardReviewOption'],
-      api_version: 'V1'
-    )
-    render_model_errors unless @notice_of_disagreement.validate
-  end
-
-  # Follows JSON API v1.0 error object standard (https://jsonapi.org/format/1.0/#error-objects)
-  def render_model_errors
-    render json: model_errors_to_json_api, status: MODEL_ERROR_STATUS
-  end
-
-  def model_errors_to_json_api
-    errors = @notice_of_disagreement.errors.map do |error|
-      data = I18n.t('common.exceptions.validation_errors').deep_merge error.options
-      data[:source] = { pointer: error.attribute.to_s }
-      data
-    end
-    { errors: errors }
-  end
-
-  def find_notice_of_disagreement
-    @id = params[:id]
-    @notice_of_disagreement = AppealsApi::NoticeOfDisagreement.find(@id)
-  rescue ActiveRecord::RecordNotFound
-    render_notice_of_disagreement_not_found
-  end
-
-  def render_notice_of_disagreement_not_found
-    render(
-      status: :not_found,
-      json: {
-        errors: [
-          { status: 404, detail: I18n.t('appeals_api.errors.nod_not_found', id: @id) }
-        ]
-      }
-    )
-  end
-
-  def render_notice_of_disagreement
-    render json: AppealsApi::NoticeOfDisagreementSerializer.new(@notice_of_disagreement).serializable_hash
+  def api_version
+    'V1'
   end
 end
